+++ processor/src/test/resources/bad_input/com/example/state/ArezOnGetterName.java
@@ -1,0 +1,26 @@
+package com.example.state;
+
+import arez.annotations.Computed;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.annotations.State;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class ArezOnGetterName
+  extends ReactArezComponent
+{
+  @Computed
+  protected abstract String getMyKey();
+
+  @State
+  protected abstract void setMyKey( String value );
+
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}